feat: add a "magic application path" for starting the flash player
Any game that has the application path ":flash:" will launch the flash
player. Since the launcher can be anywhere in relation to the flashpoint
folder (which is what the application path usually is relative to), and
because it has to be specially executed (since it wants the entry file
as the first argument, as well as the env-var "ELECTRON_RUN_AS_NODE"
removed), it only makes sense to make it a magic value.